feat: Enable stage geometry rendering with heap corruption recovery\n\n- Fix heap corruption recovery in HSD_ObjAllocAddFree\n- Save obj_heap.curr position to prevent double-allocation\n- Zero allocated memory to prevent uninitialized field crashes\n- Remove stage geometry deferral in Ground_GetStageGObj\n- Stage init chain completes, render callback registered\n- Screenshot capture works (1280x720)

+++ src/sysdolphin/baselib/objalloc.h
@@ -6,10 +6,10 @@
 #include "baselib/debug.h"
 
 typedef struct _objheap {
-    u32 top;
-    u32 curr;
-    u32 size;
-    u32 remain;
+    uintptr_t top;
+    uintptr_t curr;
+    uintptr_t size;
+    uintptr_t remain;
 } objheap;
 
 typedef struct _HSD_ObjAllocLink {
@@ -69,7 +69,7 @@ static inline void HSD_ObjAllocDisableNumLimit(HSD_ObjAllocData* data)
     data->num_limit_flag = 0;
 }
 
-void HSD_ObjSetHeap(u32 size, void* ptr);
+void HSD_ObjSetHeap(uintptr_t size, void* ptr);
 s32 HSD_ObjAllocAddFree(HSD_ObjAllocData* data, u32 num);
 void* HSD_ObjAlloc(HSD_ObjAllocData* data);
 void HSD_ObjFree(HSD_ObjAllocData* data, void* obj);
